# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: Auth
    description: Browser and mobile authentication endpoints.
  - name: Platform
    description: Public platform endpoints (feature flags, public impact, my badges).
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          description: Relative path to redirect to after login (must start with `/`). Defaults to `/`.
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to the `returnTo` path on success or `/api/login` on failure.
  /logout:
    get:
      tags: [Auth]
      operationId: beginBrowserLogout
      summary: Log out and clear session
      responses:
        '302':
          description: Redirect to the OIDC end-session endpoint.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile authorization code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExchangeMobileAuthorizationCodeBody'
      responses:
        '200':
          description: Token exchange successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExchangeMobileAuthorizationCodeResponse'
  /feature-flags:
    get:
      tags: [Platform]
      operationId: listFeatureFlags
      summary: Public map of feature flags (key → enabled).
      description: Returns enabled/disabled state for every defined feature flag. Missing keys default to enabled on the client.
      responses:
        '200':
          description: Flag map.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureFlagsResponse'
  /impact:
    get:
      tags: [Platform]
      operationId: getPublicImpact
      summary: Public impact statistics and transformation stories.
      description: Returns the headline impact numbers (with admin overrides applied) plus the published transformation stories carousel.
      responses:
        '200':
          description: Impact payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicImpactResponse'
  /my/badges:
    get:
      tags: [Platform]
      operationId: listMyBadges
      summary: Badges earned by the current user (with locked previews).
      parameters:
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Badge list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyBadgesResponse'
        '401':
          description: Not authenticated.
  /admin/feature-flags:
    get:
      tags: [Platform]
      operationId: adminListFeatureFlags
      summary: List all feature flags with descriptions (admin).
      parameters:
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Flag list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminFeatureFlagListResponse'
        '403':
          description: Forbidden.
  /admin/feature-flags/{key}:
    patch:
      tags: [Platform]
      operationId: adminUpdateFeatureFlag
      summary: Toggle a feature flag.
      parameters:
        - $ref: '#/components/parameters/SessionCookie'
        - name: key
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeatureFlagPatchBody'
      responses:
        '200':
          description: Updated flag.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminFeatureFlag'
        '400':
          description: Invalid input.
        '403':
          description: Forbidden.
        '404':
          description: Not found.
  /admin/audit-log:
    get:
      tags: [Platform]
      operationId: adminListAuditLog
      summary: Paginated audit log entries.
      parameters:
        - $ref: '#/components/parameters/SessionCookie'
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: actor
          in: query
          schema: { type: string }
        - name: entityType
          in: query
          schema: { type: string }
        - name: action
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Audit entries.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogResponse'
        '403':
          description: Forbidden.
  /admin/impact-stats:
    get:
      tags: [Platform]
      operationId: adminListImpactStats
      summary: Admin view of impact stats with override values.
      parameters:
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Impact stats.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminImpactStatsResponse'
        '403':
          description: Forbidden.
  /admin/impact-stats/{key}:
    patch:
      tags: [Platform]
      operationId: adminUpdateImpactStat
      summary: Update an impact-stat override or labels.
      parameters:
        - $ref: '#/components/parameters/SessionCookie'
        - name: key
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImpactStatPatchBody'
      responses:
        '200':
          description: Updated stat row.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminImpactStat'
        '400':
          description: Invalid input.
        '403':
          description: Forbidden.
        '404':
          description: Not found.
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Log out a mobile session
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutMobileSessionResponse'
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      schema:
        type: string
      description: Bearer token for mobile session authentication.
    SessionCookie:
      name: sid
      in: cookie
      required: false
      schema:
        type: string
      description: Session cookie for browser authentication.
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AuthUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        profileImageUrl:
          type: string
          nullable: true
      required:
        - id
    AuthUserEnvelope:
      type: object
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
      required:
        - user
    ExchangeMobileAuthorizationCodeBody:
      type: object
      properties:
        code:
          type: string
        code_verifier:
          type: string
        state:
          type: string
        nonce:
          type: string
          nullable: true
        redirect_uri:
          type: string
      required:
        - code
        - code_verifier
        - state
        - redirect_uri
    ExchangeMobileAuthorizationCodeResponse:
      type: object
      properties:
        token:
          type: string
      required:
        - token
    LogoutMobileSessionResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    FeatureFlagsResponse:
      type: object
      properties:
        flags:
          type: object
          additionalProperties:
            type: boolean
      required:
        - flags
    PublicImpactStat:
      type: object
      properties:
        key:
          type: string
        labelAr:
          type: string
        labelEn:
          type: string
        value:
          type: string
        isOverridden:
          type: boolean
      required: [key, labelAr, labelEn, value, isOverridden]
    PublicTransformationStory:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        roleAr:
          type: string
          nullable: true
        roleEn:
          type: string
          nullable: true
        quoteAr:
          type: string
        quoteEn:
          type: string
          nullable: true
        photoUrl:
          type: string
          nullable: true
        displayOrder:
          type: integer
      required: [id, name, quoteAr, displayOrder]
    PublicImpactResponse:
      type: object
      properties:
        stats:
          type: array
          items:
            $ref: '#/components/schemas/PublicImpactStat'
        stories:
          type: array
          items:
            $ref: '#/components/schemas/PublicTransformationStory'
      required: [stats, stories]
    MyBadge:
      type: object
      properties:
        key:
          type: string
        titleAr:
          type: string
        titleEn:
          type: string
        descriptionAr:
          type: string
          nullable: true
        descriptionEn:
          type: string
          nullable: true
        icon:
          type: string
        colorClass:
          type: string
        earned:
          type: boolean
        earnedAt:
          type: string
          format: date-time
          nullable: true
      required: [key, titleAr, titleEn, icon, colorClass, earned]
    AdminFeatureFlag:
      type: object
      properties:
        key: { type: string }
        enabled: { type: boolean }
        descriptionAr: { type: string, nullable: true }
        descriptionEn: { type: string, nullable: true }
        updatedAt: { type: string, format: date-time, nullable: true }
        updatedById: { type: string, nullable: true }
      required: [key, enabled]
    AdminFeatureFlagListResponse:
      type: object
      properties:
        flags:
          type: array
          items: { $ref: '#/components/schemas/AdminFeatureFlag' }
      required: [flags]
    FeatureFlagPatchBody:
      type: object
      properties:
        enabled: { type: boolean }
      required: [enabled]
    AuditLogEntry:
      type: object
      properties:
        id: { type: string }
        actorUserId: { type: string, nullable: true }
        actorEmail: { type: string, nullable: true }
        action: { type: string }
        entityType: { type: string }
        entityId: { type: string, nullable: true }
        description: { type: string, nullable: true }
        beforeJson: { nullable: true }
        afterJson: { nullable: true }
        createdAt: { type: string, format: date-time }
      required: [id, action, entityType, createdAt]
    AuditLogResponse:
      type: object
      properties:
        entries:
          type: array
          items: { $ref: '#/components/schemas/AuditLogEntry' }
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
      required: [entries, total, limit, offset]
    AdminImpactStat:
      type: object
      properties:
        key: { type: string }
        labelAr: { type: string }
        labelEn: { type: string }
        overrideValue: { type: string, nullable: true }
        realValue: { type: string }
        displayOrder: { type: integer }
      required: [key, labelAr, labelEn, realValue, displayOrder]
    AdminImpactStatsResponse:
      type: object
      properties:
        stats:
          type: array
          items: { $ref: '#/components/schemas/AdminImpactStat' }
      required: [stats]
    ImpactStatPatchBody:
      type: object
      properties:
        overrideValue: { type: string, nullable: true }
        labelAr: { type: string }
        labelEn: { type: string }
    MyBadgesResponse:
      type: object
      properties:
        badges:
          type: array
          items:
            $ref: '#/components/schemas/MyBadge'
        earnedCount:
          type: integer
        totalCount:
          type: integer
      required: [badges, earnedCount, totalCount]