# lib/api-zod/src/generated/types/adminFeatureFlag.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminFeatureFlag {
  key: string;
  enabled: boolean;
  descriptionAr?: string | null;
  descriptionEn?: string | null;
  updatedAt?: Date | null;
  updatedById?: string | null;
}