# lib/api-zod/src/generated/types/adminImpactStat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminImpactStat {
  key: string;
  labelAr: string;
  labelEn: string;
  overrideValue?: string | null;
  realValue: string;
  displayOrder: number;
}