# lib/api-zod/src/generated/types/adminImpactStatsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminImpactStat } from "./adminImpactStat";

export interface AdminImpactStatsResponse {
  stats: AdminImpactStat[];
}