# lib/api-zod/src/generated/types/auditLogEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuditLogEntry {
  id: string;
  actorUserId?: string | null;
  actorEmail?: string | null;
  action: string;
  entityType: string;
  entityId?: string | null;
  description?: string | null;
  beforeJson?: unknown | null;
  afterJson?: unknown | null;
  createdAt: Date;
}